--- crates/ruff/src/rules/pep8_naming/rules.rs
@@ -1,573 +1,0 @@
-use ruff_macros::{define_violation, derive_message_formats};
-use ruff_python::string::{self};
-use rustpython_parser::ast::{Arg, Arguments, Expr, ExprKind, Stmt};
-
-use super::helpers;
-use crate::ast::function_type;
-use crate::ast::helpers::identifier_range;
-use crate::ast::types::{Range, Scope, ScopeKind};
-use crate::checkers::ast::Checker;
-use crate::registry::Diagnostic;
-use crate::source_code::Locator;
-use crate::violation::Violation;
-
-define_violation!(
-    pub struct InvalidClassName {
-        pub name: String,
-    }
-);
-impl Violation for InvalidClassName {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        let InvalidClassName { name } = self;
-        format!("Class name `{name}` should use CapWords convention ")
-    }
-}
-
-define_violation!(
-    pub struct InvalidFunctionName {
-        pub name: String,
-    }
-);
-impl Violation for InvalidFunctionName {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        let InvalidFunctionName { name } = self;
-        format!("Function name `{name}` should be lowercase")
-    }
-}
-
-define_violation!(
-    pub struct InvalidArgumentName {
-        pub name: String,
-    }
-);
-impl Violation for InvalidArgumentName {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        let InvalidArgumentName { name } = self;
-        format!("Argument name `{name}` should be lowercase")
-    }
-}
-
-define_violation!(
-    pub struct InvalidFirstArgumentNameForClassMethod;
-);
-impl Violation for InvalidFirstArgumentNameForClassMethod {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        format!("First argument of a class method should be named `cls`")
-    }
-}
-
-define_violation!(
-    pub struct InvalidFirstArgumentNameForMethod;
-);
-impl Violation for InvalidFirstArgumentNameForMethod {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        format!("First argument of a method should be named `self`")
-    }
-}
-
-define_violation!(
-    pub struct NonLowercaseVariableInFunction {
-        pub name: String,
-    }
-);
-impl Violation for NonLowercaseVariableInFunction {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        let NonLowercaseVariableInFunction { name } = self;
-        format!("Variable `{name}` in function should be lowercase")
-    }
-}
-
-define_violation!(
-    pub struct DunderFunctionName;
-);
-impl Violation for DunderFunctionName {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        format!("Function name should not start and end with `__`")
-    }
-}
-
-define_violation!(
-    pub struct ConstantImportedAsNonConstant {
-        pub name: String,
-        pub asname: String,
-    }
-);
-impl Violation for ConstantImportedAsNonConstant {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        let ConstantImportedAsNonConstant { name, asname } = self;
-        format!("Constant `{name}` imported as non-constant `{asname}`")
-    }
-}
-
-define_violation!(
-    pub struct LowercaseImportedAsNonLowercase {
-        pub name: String,
-        pub asname: String,
-    }
-);
-impl Violation for LowercaseImportedAsNonLowercase {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        let LowercaseImportedAsNonLowercase { name, asname } = self;
-        format!("Lowercase `{name}` imported as non-lowercase `{asname}`")
-    }
-}
-
-define_violation!(
-    pub struct CamelcaseImportedAsLowercase {
-        pub name: String,
-        pub asname: String,
-    }
-);
-impl Violation for CamelcaseImportedAsLowercase {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        let CamelcaseImportedAsLowercase { name, asname } = self;
-        format!("Camelcase `{name}` imported as lowercase `{asname}`")
-    }
-}
-
-define_violation!(
-    pub struct CamelcaseImportedAsConstant {
-        pub name: String,
-        pub asname: String,
-    }
-);
-impl Violation for CamelcaseImportedAsConstant {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        let CamelcaseImportedAsConstant { name, asname } = self;
-        format!("Camelcase `{name}` imported as constant `{asname}`")
-    }
-}
-
-define_violation!(
-    pub struct MixedCaseVariableInClassScope {
-        pub name: String,
-    }
-);
-impl Violation for MixedCaseVariableInClassScope {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        let MixedCaseVariableInClassScope { name } = self;
-        format!("Variable `{name}` in class scope should not be mixedCase")
-    }
-}
-
-define_violation!(
-    pub struct MixedCaseVariableInGlobalScope {
-        pub name: String,
-    }
-);
-impl Violation for MixedCaseVariableInGlobalScope {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        let MixedCaseVariableInGlobalScope { name } = self;
-        format!("Variable `{name}` in global scope should not be mixedCase")
-    }
-}
-
-define_violation!(
-    pub struct CamelcaseImportedAsAcronym {
-        pub name: String,
-        pub asname: String,
-    }
-);
-impl Violation for CamelcaseImportedAsAcronym {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        let CamelcaseImportedAsAcronym { name, asname } = self;
-        format!("Camelcase `{name}` imported as acronym `{asname}`")
-    }
-}
-
-define_violation!(
-    pub struct ErrorSuffixOnExceptionName {
-        pub name: String,
-    }
-);
-impl Violation for ErrorSuffixOnExceptionName {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        let ErrorSuffixOnExceptionName { name } = self;
-        format!("Exception name `{name}` should be named with an Error suffix")
-    }
-}
-
-/// N801
-pub fn invalid_class_name(class_def: &Stmt, name: &str, locator: &Locator) -> Option<Diagnostic> {
-    let stripped = name.strip_prefix('_').unwrap_or(name);
-    if !stripped.chars().next().map_or(false, char::is_uppercase) || stripped.contains('_') {
-        return Some(Diagnostic::new(
-            InvalidClassName {
-                name: name.to_string(),
-            },
-            identifier_range(class_def, locator),
-        ));
-    }
-    None
-}
-
-/// N802
-pub fn invalid_function_name(
-    func_def: &Stmt,
-    name: &str,
-    ignore_names: &[String],
-    locator: &Locator,
-) -> Option<Diagnostic> {
-    if ignore_names.iter().any(|ignore_name| ignore_name == name) {
-        return None;
-    }
-    if name.to_lowercase() != name {
-        return Some(Diagnostic::new(
-            InvalidFunctionName {
-                name: name.to_string(),
-            },
-            identifier_range(func_def, locator),
-        ));
-    }
-    None
-}
-
-/// N803
-pub fn invalid_argument_name(name: &str, arg: &Arg, ignore_names: &[String]) -> Option<Diagnostic> {
-    if ignore_names.iter().any(|ignore_name| ignore_name == name) {
-        return None;
-    }
-    if name.to_lowercase() != name {
-        return Some(Diagnostic::new(
-            InvalidArgumentName {
-                name: name.to_string(),
-            },
-            Range::from_located(arg),
-        ));
-    }
-    None
-}
-
-/// N804
-pub fn invalid_first_argument_name_for_class_method(
-    checker: &Checker,
-    scope: &Scope,
-    name: &str,
-    decorator_list: &[Expr],
-    args: &Arguments,
-) -> Option<Diagnostic> {
-    if !matches!(
-        function_type::classify(
-            checker,
-            scope,
-            name,
-            decorator_list,
-            &checker.settings.pep8_naming.classmethod_decorators,
-            &checker.settings.pep8_naming.staticmethod_decorators,
-        ),
-        function_type::FunctionType::ClassMethod
-    ) {
-        return None;
-    }
-    if let Some(arg) = args.posonlyargs.first().or_else(|| args.args.first()) {
-        if arg.node.arg != "cls" {
-            if checker
-                .settings
-                .pep8_naming
-                .ignore_names
-                .iter()
-                .any(|ignore_name| ignore_name == name)
-            {
-                return None;
-            }
-            return Some(Diagnostic::new(
-                InvalidFirstArgumentNameForClassMethod,
-                Range::from_located(arg),
-            ));
-        }
-    }
-    None
-}
-
-/// N805
-pub fn invalid_first_argument_name_for_method(
-    checker: &Checker,
-    scope: &Scope,
-    name: &str,
-    decorator_list: &[Expr],
-    args: &Arguments,
-) -> Option<Diagnostic> {
-    if !matches!(
-        function_type::classify(
-            checker,
-            scope,
-            name,
-            decorator_list,
-            &checker.settings.pep8_naming.classmethod_decorators,
-            &checker.settings.pep8_naming.staticmethod_decorators,
-        ),
-        function_type::FunctionType::Method
-    ) {
-        return None;
-    }
-    let arg = args.posonlyargs.first().or_else(|| args.args.first())?;
-    if arg.node.arg == "self" {
-        return None;
-    }
-    if checker
-        .settings
-        .pep8_naming
-        .ignore_names
-        .iter()
-        .any(|ignore_name| ignore_name == name)
-    {
-        return None;
-    }
-    Some(Diagnostic::new(
-        InvalidFirstArgumentNameForMethod,
-        Range::from_located(arg),
-    ))
-}
-
-/// N806
-pub fn non_lowercase_variable_in_function(
-    checker: &mut Checker,
-    expr: &Expr,
-    stmt: &Stmt,
-    name: &str,
-) {
-    if checker
-        .settings
-        .pep8_naming
-        .ignore_names
-        .iter()
-        .any(|ignore_name| ignore_name == name)
-    {
-        return;
-    }
-
-    if name.to_lowercase() != name
-        && !helpers::is_namedtuple_assignment(checker, stmt)
-        && !helpers::is_typeddict_assignment(checker, stmt)
-        && !helpers::is_type_var_assignment(checker, stmt)
-    {
-        checker.diagnostics.push(Diagnostic::new(
-            NonLowercaseVariableInFunction {
-                name: name.to_string(),
-            },
-            Range::from_located(expr),
-        ));
-    }
-}
-
-/// N807
-pub fn dunder_function_name(
-    scope: &Scope,
-    stmt: &Stmt,
-    name: &str,
-    locator: &Locator,
-) -> Option<Diagnostic> {
-    if matches!(scope.kind, ScopeKind::Class(_)) {
-        return None;
-    }
-    if !(name.starts_with("__") && name.ends_with("__")) {
-        return None;
-    }
-    // Allowed under PEP 562 (https://peps.python.org/pep-0562/).
-    if matches!(scope.kind, ScopeKind::Module) && (name == "__getattr__" || name == "__dir__") {
-        return None;
-    }
-
-    Some(Diagnostic::new(
-        DunderFunctionName,
-        identifier_range(stmt, locator),
-    ))
-}
-
-/// N811
-pub fn constant_imported_as_non_constant(
-    import_from: &Stmt,
-    name: &str,
-    asname: &str,
-    locator: &Locator,
-) -> Option<Diagnostic> {
-    if string::is_upper(name) && !string::is_upper(asname) {
-        return Some(Diagnostic::new(
-            ConstantImportedAsNonConstant {
-                name: name.to_string(),
-                asname: asname.to_string(),
-            },
-            identifier_range(import_from, locator),
-        ));
-    }
-    None
-}
-
-/// N812
-pub fn lowercase_imported_as_non_lowercase(
-    import_from: &Stmt,
-    name: &str,
-    asname: &str,
-    locator: &Locator,
-) -> Option<Diagnostic> {
-    if !string::is_upper(name) && string::is_lower(name) && asname.to_lowercase() != asname {
-        return Some(Diagnostic::new(
-            LowercaseImportedAsNonLowercase {
-                name: name.to_string(),
-                asname: asname.to_string(),
-            },
-            identifier_range(import_from, locator),
-        ));
-    }
-    None
-}
-
-/// N813
-pub fn camelcase_imported_as_lowercase(
-    import_from: &Stmt,
-    name: &str,
-    asname: &str,
-    locator: &Locator,
-) -> Option<Diagnostic> {
-    if helpers::is_camelcase(name) && string::is_lower(asname) {
-        return Some(Diagnostic::new(
-            CamelcaseImportedAsLowercase {
-                name: name.to_string(),
-                asname: asname.to_string(),
-            },
-            identifier_range(import_from, locator),
-        ));
-    }
-    None
-}
-
-/// N814
-pub fn camelcase_imported_as_constant(
-    import_from: &Stmt,
-    name: &str,
-    asname: &str,
-    locator: &Locator,
-) -> Option<Diagnostic> {
-    if helpers::is_camelcase(name)
-        && !string::is_lower(asname)
-        && string::is_upper(asname)
-        && !helpers::is_acronym(name, asname)
-    {
-        return Some(Diagnostic::new(
-            CamelcaseImportedAsConstant {
-                name: name.to_string(),
-                asname: asname.to_string(),
-            },
-            identifier_range(import_from, locator),
-        ));
-    }
-    None
-}
-
-/// N815
-pub fn mixed_case_variable_in_class_scope(
-    checker: &mut Checker,
-    expr: &Expr,
-    stmt: &Stmt,
-    name: &str,
-) {
-    if checker
-        .settings
-        .pep8_naming
-        .ignore_names
-        .iter()
-        .any(|ignore_name| ignore_name == name)
-    {
-        return;
-    }
-    if helpers::is_mixed_case(name) && !helpers::is_namedtuple_assignment(checker, stmt) {
-        checker.diagnostics.push(Diagnostic::new(
-            MixedCaseVariableInClassScope {
-                name: name.to_string(),
-            },
-            Range::from_located(expr),
-        ));
-    }
-}
-
-/// N816
-pub fn mixed_case_variable_in_global_scope(
-    checker: &mut Checker,
-    expr: &Expr,
-    stmt: &Stmt,
-    name: &str,
-) {
-    if checker
-        .settings
-        .pep8_naming
-        .ignore_names
-        .iter()
-        .any(|ignore_name| ignore_name == name)
-    {
-        return;
-    }
-    if helpers::is_mixed_case(name) && !helpers::is_namedtuple_assignment(checker, stmt) {
-        checker.diagnostics.push(Diagnostic::new(
-            MixedCaseVariableInGlobalScope {
-                name: name.to_string(),
-            },
-            Range::from_located(expr),
-        ));
-    }
-}
-
-/// N817
-pub fn camelcase_imported_as_acronym(
-    import_from: &Stmt,
-    name: &str,
-    asname: &str,
-    locator: &Locator,
-) -> Option<Diagnostic> {
-    if helpers::is_camelcase(name)
-        && !string::is_lower(asname)
-        && string::is_upper(asname)
-        && helpers::is_acronym(name, asname)
-    {
-        return Some(Diagnostic::new(
-            CamelcaseImportedAsAcronym {
-                name: name.to_string(),
-                asname: asname.to_string(),
-            },
-            identifier_range(import_from, locator),
-        ));
-    }
-    None
-}
-
-/// N818
-pub fn error_suffix_on_exception_name(
-    class_def: &Stmt,
-    bases: &[Expr],
-    name: &str,
-    locator: &Locator,
-) -> Option<Diagnostic> {
-    if !bases.iter().any(|base| {
-        if let ExprKind::Name { id, .. } = &base.node {
-            id == "Exception" || id.ends_with("Error")
-        } else {
-            false
-        }
-    }) {
-        return None;
-    }
-
-    if name.ends_with("Error") {
-        return None;
-    }
-    Some(Diagnostic::new(
-        ErrorSuffixOnExceptionName {
-            name: name.to_string(),
-        },
-        identifier_range(class_def, locator),
-    ))
-}
