+++ crates/ruff/src/rules/pep8_naming/rules/invalid_first_argument_name_for_method.rs
@@ -1,0 +1,91 @@
+use ruff_macros::{define_violation, derive_message_formats};
+use rustpython_parser::ast::{Arguments, Expr};
+
+use crate::ast::function_type;
+use crate::ast::types::{Range, Scope};
+use crate::checkers::ast::Checker;
+use crate::registry::Diagnostic;
+use crate::violation::Violation;
+
+define_violation!(
+    /// ## What it does
+    /// Checks for instance methods to have `self` as the first argument.
+    ///
+    /// ## Why is this bad?
+    /// Using `self` as first argument for instance methods is recommended by [PEP8]:
+    ///
+    /// > Always use self for the first argument to instance methods.
+    /// >
+    /// > If a function argument’s name clashes with a reserved keyword, it is generally better to
+    /// > append a single trailing underscore rather than use an abbreviation or spelling corruption.
+    /// > Thus class_ is better than clss. (Perhaps better is to avoid such clashes by using a synonym.)
+    ///
+    /// ## Options
+    /// * `pep8-naming.classmethod-decorators`
+    /// * `pep8-naming.staticmethod-decorators`
+    /// * `pep8-naming.ignore-names`
+    ///
+    /// ## Example
+    /// ```python
+    /// class Example:
+    ///     def function(cls, data):
+    ///         ...
+    /// ```
+    ///
+    /// Use instead:
+    /// ```python
+    /// class Example:
+    ///     def function(self, data):
+    ///         ...
+    /// ```
+    ///
+    /// [PEP8]: https://peps.python.org/pep-0008/#function-and-method-arguments
+    ///
+    pub struct InvalidFirstArgumentNameForMethod;
+);
+impl Violation for InvalidFirstArgumentNameForMethod {
+    #[derive_message_formats]
+    fn message(&self) -> String {
+        format!("First argument of a method should be named `self`")
+    }
+}
+
+/// N805
+pub fn invalid_first_argument_name_for_method(
+    checker: &Checker,
+    scope: &Scope,
+    name: &str,
+    decorator_list: &[Expr],
+    args: &Arguments,
+) -> Option<Diagnostic> {
+    if !matches!(
+        function_type::classify(
+            checker,
+            scope,
+            name,
+            decorator_list,
+            &checker.settings.pep8_naming.classmethod_decorators,
+            &checker.settings.pep8_naming.staticmethod_decorators,
+        ),
+        function_type::FunctionType::Method
+    ) {
+        return None;
+    }
+    let arg = args.posonlyargs.first().or_else(|| args.args.first())?;
+    if arg.node.arg == "self" {
+        return None;
+    }
+    if checker
+        .settings
+        .pep8_naming
+        .ignore_names
+        .iter()
+        .any(|ignore_name| ignore_name == name)
+    {
+        return None;
+    }
+    Some(Diagnostic::new(
+        InvalidFirstArgumentNameForMethod,
+        Range::from_located(arg),
+    ))
+}
