+++ crates/ruff/src/rules/pep8_naming/rules/invalid_function_name.rs
@@ -1,0 +1,68 @@
+use ruff_macros::{define_violation, derive_message_formats};
+use rustpython_parser::ast::Stmt;
+
+use crate::ast::helpers::identifier_range;
+use crate::registry::Diagnostic;
+use crate::source_code::Locator;
+use crate::violation::Violation;
+
+define_violation!(
+    /// ## What it does
+    /// Checks for functions names that are following the [`snake_case`] naming convention.
+    ///
+    /// ## Why is this bad?
+    /// [PEP8] recommends that function names follow `snake_case`:
+    ///
+    /// > Function names should be lowercase, with words separated by underscores as necessary to
+    /// > improve readability. mixedCase is allowed only in contexts where that’s already the
+    /// > prevailing style (e.g. threading.py), to retain backwards compatibility.
+    ///
+    /// ## Options
+    /// * `pep8-naming.ignore-names`
+    ///
+    /// ## Example
+    /// ```python
+    /// def myFunction():
+    ///     pass
+    /// ```
+    ///
+    /// Use instead:
+    /// ```python
+    /// def my_function():
+    ///     pass
+    /// ```
+    ///
+    /// [PEP8]: https://peps.python.org/pep-0008/#function-and-variable-names
+    /// [`snake_case`]: https://en.wikipedia.org/wiki/Snake_case
+    pub struct InvalidFunctionName {
+        pub name: String,
+    }
+);
+impl Violation for InvalidFunctionName {
+    #[derive_message_formats]
+    fn message(&self) -> String {
+        let InvalidFunctionName { name } = self;
+        format!("Function name `{name}` should be lowercase")
+    }
+}
+
+/// N802
+pub fn invalid_function_name(
+    func_def: &Stmt,
+    name: &str,
+    ignore_names: &[String],
+    locator: &Locator,
+) -> Option<Diagnostic> {
+    if ignore_names.iter().any(|ignore_name| ignore_name == name) {
+        return None;
+    }
+    if name.to_lowercase() != name {
+        return Some(Diagnostic::new(
+            InvalidFunctionName {
+                name: name.to_string(),
+            },
+            identifier_range(func_def, locator),
+        ));
+    }
+    None
+}
