+++ crates/ruff/src/rules/pep8_naming/rules/non_lowercase_variable_in_function.rs
@@ -1,0 +1,82 @@
+use ruff_macros::{define_violation, derive_message_formats};
+use rustpython_parser::ast::{Expr, Stmt};
+
+use crate::ast::types::Range;
+use crate::checkers::ast::Checker;
+use crate::registry::Diagnostic;
+use crate::rules::pep8_naming::helpers;
+use crate::violation::Violation;
+
+define_violation!(
+    /// ## What it does
+    /// Checks for non-lowercase variables in functions.
+    ///
+    /// ## Why is this bad?
+    /// Variable names in functions should be lowercase as recommended by [PEP8]:
+    ///
+    /// > Function names should be lowercase, with words separated by underscores as necessary to
+    /// > improve readability. Variable names follow the same convention as function names. mixedCase
+    /// > is allowed only in contexts where that's already the prevailing style (e.g. threading.py),
+    /// to retain backwards compatibility.
+    ///
+    /// ## Options
+    /// * `pep8-naming.ignore-names`
+    ///
+    /// ## Example
+    /// ```python
+    /// def my_function(a):
+    ///     B = a + 3
+    ///     return B
+    /// ```
+    ///
+    /// Use instead:
+    /// ```python
+    /// def my_function(a):
+    ///     b = a + 3
+    ///     return b
+    /// ```
+    ///
+    /// [PEP8]: https://peps.python.org/pep-0008/#function-and-variable-names
+    ///
+    pub struct NonLowercaseVariableInFunction {
+        pub name: String,
+    }
+);
+impl Violation for NonLowercaseVariableInFunction {
+    #[derive_message_formats]
+    fn message(&self) -> String {
+        let NonLowercaseVariableInFunction { name } = self;
+        format!("Variable `{name}` in function should be lowercase")
+    }
+}
+
+/// N806
+pub fn non_lowercase_variable_in_function(
+    checker: &mut Checker,
+    expr: &Expr,
+    stmt: &Stmt,
+    name: &str,
+) {
+    if checker
+        .settings
+        .pep8_naming
+        .ignore_names
+        .iter()
+        .any(|ignore_name| ignore_name == name)
+    {
+        return;
+    }
+
+    if name.to_lowercase() != name
+        && !helpers::is_namedtuple_assignment(checker, stmt)
+        && !helpers::is_typeddict_assignment(checker, stmt)
+        && !helpers::is_type_var_assignment(checker, stmt)
+    {
+        checker.diagnostics.push(Diagnostic::new(
+            NonLowercaseVariableInFunction {
+                name: name.to_string(),
+            },
+            Range::from_located(expr),
+        ));
+    }
+}
