+++ crates/ruff_python/src/string.rs
@@ -3,7 +3,8 @@ use regex::Regex;
 
 pub static STRING_QUOTE_PREFIX_REGEX: Lazy<Regex> =
     Lazy::new(|| Regex::new(r#"^(?i)[urb]*['"](?P<raw>.*)['"]$"#).unwrap());
-pub static LOWER_OR_UNDERSCORE: Lazy<Regex> = Lazy::new(|| Regex::new(r"^[a-z_]+$").unwrap());
+pub static LOWER_OR_UNDERSCORE: Lazy<Regex> =
+    Lazy::new(|| Regex::new(r"^[a-z][a-z0-9_]*$").unwrap());
 
 pub fn is_lower(s: &str) -> bool {
     let mut cased = false;
@@ -64,10 +65,15 @@ mod tests {
 
     #[test]
     fn test_is_lower_underscore() {
+        assert!(is_lower_with_underscore("a"));
         assert!(is_lower_with_underscore("abc"));
+        assert!(is_lower_with_underscore("abc0"));
+        assert!(is_lower_with_underscore("abc_"));
         assert!(is_lower_with_underscore("a_b_c"));
         assert!(!is_lower_with_underscore("a-b-c"));
         assert!(!is_lower_with_underscore("a_B_c"));
+        assert!(!is_lower_with_underscore("0abc"));
+        assert!(!is_lower_with_underscore("_abc"));
     }
 
     #[test]
