+++ ruff.schema.json
@@ -1656,6 +1656,9 @@
         "N816",
         "N817",
         "N818",
+        "N9",
+        "N99",
+        "N999",
         "NPY",
         "NPY0",
         "NPY00",
